+++ xash3d_engine/engine/client/s_mouth.c
@@ -67,7 +67,7 @@ void SND_MoveMouth8(channel_t* ch, wavdata_t* pSource, int count)
 	scount = pMouth->sndcount;
 	savg = 0;
 
-	while ( i < count && scount < CAVGSAMPLES )
+	while ( i < (uint)count && scount < CAVGSAMPLES )
 	{
 		data = pdata[i];
 		savg += abs(data);
@@ -118,7 +118,7 @@ void SND_MoveMouth16(channel_t* ch, wavdata_t* pSource, int count)
 	scount = pMouth->sndcount;
 	savg = 0;
 
-	while ( i < count && scount < CAVGSAMPLES )
+	while ( i < (uint)count && scount < CAVGSAMPLES )
 	{
 		data = pdata[i];
 		data = (bound(-32767, data, 0x7ffe) >> 8);
@@ -184,7 +184,7 @@ void SND_MoveMouthRaw(rawchan_t* ch, portable_samplepair_t* pData, int count)
 	scount = pMouth->sndcount;
 	savg = 0;
 
-	while ( i < count && scount < CAVGSAMPLES )
+	while ( i < (uint)count && scount < CAVGSAMPLES )
 	{
 		data = pData[i].left;  // mono sound anyway
 		data = (bound(-32767, data, 0x7ffe) >> 8);
